backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
PyMySQL
python-dotenv
Werkzeug
Flask-MySQLdb